Add refreshToken
- add db table
- add refresh logic on new endpoint

+++ src/Domain/Common/Options/JwtOptions.cs
@@ -5,7 +5,7 @@ namespace SmartHub.Domain.Common.Options
 	public class JwtOptions
 	{
 		[Required] public string Key { get; set; } = default!;
-		public double LifeTimeInMinutes { get; set; }
+		public int LifeTimeInMinutes { get; set; }
 		public string? Issuer { get; set; }
 		public string? Audience { get; set; }
 	}
